requirements.txt
playwright
pytest-playwright
python-dotenv